replacing and renaming builtin actions package to fix linter warnings
